Revamp drawScale due to new log scale requirements. Make the scale including labels better partially refreshable.

+++ main/Colors.h
@@ -1,13 +1,13 @@
 #pragma once
 
-extern int g_col_background;
-extern int g_col_highlight;
-extern int g_col_header_r;
-extern int g_col_header_g;
-extern int g_col_header_b;
-extern int g_col_header_light_r;
-extern int g_col_header_light_g;
-extern int g_col_header_light_b;
+extern uint8_t g_col_background;
+extern uint8_t g_col_highlight;
+extern uint8_t g_col_header_r;
+extern uint8_t g_col_header_g;
+extern uint8_t g_col_header_b;
+extern uint8_t g_col_header_light_r;
+extern uint8_t g_col_header_light_g;
+extern uint8_t g_col_header_light_b;
 
 // use extern int g_col_highlight and g_col_backgroud to define black and white
 // use xetren int g_col_header_xxx to define header color
